search address
This function takes three parameters area, zipcode, and sector. It checks which parameter is provided and adds the corresponding filter to a query set for the Address model.

If area is provided, it filters the addresses based on a case-insensitive match of the area field. If zipcode is provided, it filters the addresses based on a case-insensitive match of the zip_code field. If sector is provided, it filters the addresses based on a case-insensitive match of the sector field or the subsector field.

If none of the input parameters are provided, the function returns an empty query set.